fix: move theme backgrounds to helpers and remove repeated call
