+++ scripts/helpers/themeBackgrounds.js
